+++ lib/src/utils/utils.dart
@@ -7,6 +7,7 @@ import 'package:keep/src/keep.dart';
 import 'package:keep/src/key/key.dart';
 import 'package:keep/src/storage/storage.dart';
 
+part 'type.dart';
 part 'codec.dart';
 part 'exception.dart';
 part 'widgets.dart';